+++ index.js
@@ -35,6 +35,8 @@ var pathLoader = require('path-loader');
 var qs = require('querystring');
 var URI = require('uri-js');
 
+var ptrTokenRegex = /[~].?/g;
+var ptrTokens = ['0', '1'];
 var remoteCache = {};
 var remoteTypes = ['relative', 'remote'];
 var uriDetailsCache = {};
@@ -669,6 +671,7 @@ function getRefDetails (obj) {
 function isPtr (ptr) {
   var valid = isType(ptr, 'String');
   var firstChar;
+  var tokens;
 
   if (valid) {
     if (ptr !== '') {
@@ -678,6 +681,14 @@ function isPtr (ptr) {
         valid = false;
       } else if (firstChar === '#' && ptr !== '#' && ptr.charAt(1) !== '/') {
         valid = false;
+      } else {
+        tokens = ptr.match(ptrTokenRegex);
+
+        if (isType(tokens, 'Array')) {
+          valid = tokens.every(function (token) {
+            return ptrTokens.indexOf(token.split('~')[1]) > -1;
+          });
+        }
       }
     }
   }